categories

+++ src/QUI/ERP/Products/Category/Category.php
@@ -26,13 +26,35 @@ class Category extends QUI\QDOM
     protected $id;
 
     /**
+     * Parent-ID
+     *
+     * @var integer
+     */
+    protected $parentId;
+
+    /**
+     * @var
+     */
+    protected $fields;
+
+    /**
      * Modell constructor.
      *
      * @param integer $categoryId
+     * @param array $data - optional, category data
      */
-    public function __construct($categoryId)
+    public function __construct($categoryId, $data)
     {
-        $this->id = (int)$categoryId;
+        $this->parentId = 0;
+        $this->id       = (int)$categoryId;
+
+        if (isset($data['fields'])) {
+
+        }
+
+        if (isset($data['parentId'])) {
+            $this->parentId = (int)$data['parentId'];
+        }
 
         if (defined('QUIQQER_BACKEND')) {
             $this->setAttribute('viewType', 'backend');
@@ -45,6 +67,105 @@ class Category extends QUI\QDOM
     protected function getController()
     {
         return new Controller($this);
+    }
+
+    /**
+     * Return the title / name of the category
+     *
+     * @param QUI\Locale|Boolean $Locale - optional
+     * @return string
+     */
+    public function getTitle($Locale = false)
+    {
+        if (!$Locale) {
+            return QUI::getLocale()->get(
+                'quiqqer/products',
+                'products.category.' . $this->getId() . '.title'
+            );
+        }
+
+        return $Locale->get(
+            'quiqqer/products',
+            'products.category.' . $this->getId() . '.title'
+        );
+    }
+
+    /**
+     * Return the title / name of the category
+     *
+     * @param QUI\Locale|Boolean $Locale - optional
+     * @return string
+     */
+    public function getDescription($Locale = false)
+    {
+        if (!$Locale) {
+            return QUI::getLocale()->get(
+                'quiqqer/products',
+                'products.category.' . $this->getId() . '.description'
+            );
+        }
+
+        return $Locale->get(
+            'quiqqer/products',
+            'products.category.' . $this->getId() . '.description'
+        );
+    }
+
+    /**
+     * Return Field-ID
+     *
+     * @return int
+     */
+    public function getId()
+    {
+        return $this->id;
+    }
+
+    /**
+     * Return the Id of the parent category
+     * Category 0 has no parent => returns false
+     *
+     * @return integer|boolean
+     */
+    public function getParentId()
+    {
+        if ($this->getId() === 0) {
+            return false;
+        }
+
+        return $this->parentId;
+    }
+
+    /**
+     * Return the the parent category
+     * Category 0 has no parent => returns false
+     *
+     * @return bool|Category
+     * @throws QUI\Exception
+     */
+    public function getParent()
+    {
+        if ($this->getId() === 0) {
+            return false;
+        }
+
+        return QUI\ERP\Products\Handler\Categories::getCategory($this->parentId);
+    }
+
+    /**
+     * Return the attributes
+     * @return array
+     */
+    public function getAttributes()
+    {
+        $attributes       = parent::getAttributes();
+        $attributes['id'] = $this->getId();
+
+        $attributes['title']         = $this->getTitle();
+        $attributes['description']   = $this->getDescription();
+        $attributes['countChildren'] = $this->countChildren();
+
+        return $attributes;
     }
 
     /**
@@ -78,13 +199,53 @@ class Category extends QUI\QDOM
     }
 
     /**
-     * Return Field-ID
-     *
-     * @return int
+     * @return array
      */
-    public function getId()
+    public function getChildren()
     {
-        return $this->id;
+        return QUI\ERP\Products\Handler\Categories::getCategories(array(
+            'where' => array(
+                'parentId' => $this->getId()
+            )
+        ));
+    }
+
+    /**
+     * Return the number of the children
+     *
+     * @return integer
+     */
+    public function countChildren()
+    {
+        $data = QUI::getDataBase()->fetch(array(
+            'from' => QUI\ERP\Products\Utils\Tables::getCategoryTableName(),
+            'count' => array(
+                'select' => 'id',
+                'as' => 'id'
+            ),
+            'where' => array(
+                'parentId' => $this->getId()
+            )
+        ));
+
+        if (isset($data[0]) && isset($data[0]['id'])) {
+            return (int)$data[0]['id'];
+        }
+
+        return 0;
+    }
+
+    /**
+     * Return the category site
+     *
+     * @param QUI\Projects\Project $Project
+     * @return QUI\Projects\Site
+     *
+     * @throws QUI\Exception
+     */
+    public function getSite(QUI\Projects\Project $Project)
+    {
+
     }
 
     /**
@@ -100,6 +261,10 @@ class Category extends QUI\QDOM
      */
     public function delete()
     {
+        if ($this->getId() === 0) {
+            return;
+        }
+
         $this->getController()->delete();
     }
 }
